backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml>=5.3.0

# Data Validation
pydantic>=2.9.0

# LangChain and LLM
langchain-core==0.1.16
langchain-community==0.0.16
langchain-google-genai==0.0.6

# Environment Variables
python-dotenv==1.0.0
